+++ TimmyTools.WpfUi/Models/BreakTimerSettingsModel.cs
@@ -4,4 +4,5 @@ public class BreakTimerSettingsModel : BaseModel
 {
     public string ClassTitle { get; set => SetProperty(ref field, value); } = "";
     public string NextUp { get; set => SetProperty(ref field, value); } = "";
+    public int Segment { get; set => SetProperty(ref field, value); } = 0;
 }